Some packages used to be included in the standard library but will be removed in versions 3.4 or 3.5 of Ruby
Example output with a basic test deck:

[...]/ruby/3.3.0/gems/squib-0.20.0/lib/squib.rb:7: warning: logger was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add logger to your Gemfile or gemspec to silence this warning.

[...]/ruby/3.3.0/gems/squib-0.20.0/lib/squib.rb:7: warning: abbrev was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.4.0.
You can add abbrev to your Gemfile or gemspec to silence this warning.

[...]/ruby/3.3.0/gems/squib-0.20.0/lib/squib.rb:7: warning: bigdecimal was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.4.0.
You can add bigdecimal to your Gemfile or gemspec to silence this warning.

[...]/ruby/3.3.0/gems/squib-0.20.0/lib/squib.rb:8: warning: ostruct was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add ostruct to your Gemfile or gemspec to silence this warning.

[...]/ruby/3.3.0/gems/squib-0.20.0/lib/squib.rb:8: warning: csv was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.4.0.
You can add csv to your Gemfile or gemspec to silence this warning.